Move Incubator.Dialog to Dialog
