+++ src/components/picker/index.tsx
@@ -30,7 +30,7 @@ import {
   PickerItemsListProps,
   PickerMethods
 } from './types';
-import {DialogProps} from '../../incubator/dialog';
+import {DialogProps} from '../dialog';
 
 const DEFAULT_DIALOG_PROPS: DialogProps = {
   bottom: true,
