+++ src/incubator/expandableOverlay/index.tsx
@@ -3,7 +3,7 @@ import React, {useCallback, useState, forwardRef, PropsWithChildren, useImperati
 import TouchableOpacity, {TouchableOpacityProps} from '../../components/touchableOpacity';
 import View from '../../components/view';
 import Modal, {ModalProps, ModalTopBarProps} from '../../components/modal';
-import Dialog, {DialogProps} from '../dialog';
+import Dialog, {DialogProps} from '../../components/dialog';
 import {Colors} from 'style';
 
 export interface ExpandableOverlayMethods {
