+++ src/index.ts
@@ -64,7 +64,15 @@ export {default as ColorSwatch, ColorSwatchProps, ColorInfo} from './components/
 export {default as ConnectionStatusBar, ConnectionStatusBarProps} from './components/connectionStatusBar';
 export {default as Dash, DashProps} from './components/dash';
 export {default as DateTimePicker, DateTimePickerProps, DateTimePickerMode} from './components/dateTimePicker';
-export {default as Dialog, DialogProps, DialogDirections, DialogDirectionsEnum} from './incubator/dialog';
+export {
+  default as Dialog,
+  DialogProps,
+  DialogHeaderProps,
+  DialogStatics,
+  DialogImperativeMethods,
+  DialogDirections,
+  DialogDirectionsEnum
+} from './components/dialog';
 export {default as Drawer, DrawerProps, DrawerItemProps} from './components/drawer';
 export {default as ExpandableSection, ExpandableSectionProps} from './components/expandableSection';
 export {default as Fader, FaderProps, FaderPosition} from './components/fader';
